Fix bug in find_next_newline with CRLF newlines
We were incorrectly returning the location of the '\r' instead of the
'\n' in this case

+++ src/vroom_big_int.h
@@ -9,7 +9,9 @@
 #include "vroom.h"
 
 namespace cpp11 {
-inline bool is_na(long long x) { return x == NA_INTEGER64; }
+inline bool is_na(long long x) {
+  return static_cast<unsigned long>(x) == NA_INTEGER64;
+}
 } // namespace cpp11
 
 namespace vroom {
